backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
SQLAlchemy==2.0.48
psycopg[binary]==3.3.3
python-dotenv==1.2.2
PyJWT==2.12.0
pwdlib[argon2]==0.3.0
email-validator==2.3.0
alembic==1.18.4
httpx==0.28.1
pytest==9.0.2
